+++ src/main/java/de/maxhenkel/car/entity/car/parts/PartEngine3Cylinder.java
@@ -4,10 +4,13 @@ import de.maxhenkel.car.Config;
 import de.maxhenkel.car.sounds.ModSounds;
 import net.minecraft.util.SoundEvent;
 
-public class PartEngine3Cylinder extends PartEngine{
+public class PartEngine3Cylinder extends PartEngine {
 
     public PartEngine3Cylinder() {
-        super(Config.engine3CylinderMaxSpeed, Config.engine3CylinderMaxReverseSpeed, Config.engine3CylinderAcceleration, Config.engine3CylinderFuelEfficiency);
+        this.maxSpeed = () -> Config.engine3CylinderMaxSpeed.get().floatValue();
+        this.maxReverseSpeed = () -> Config.engine3CylinderMaxReverseSpeed.get().floatValue();
+        this.acceleration = () -> Config.engine3CylinderAcceleration.get().floatValue();
+        this.fuelEfficiency = () -> Config.engine3CylinderFuelEfficiency.get().floatValue();
     }
 
     @Override