+++ src/main/java/de/maxhenkel/tools/FloatSupplier.java
@@ -1,0 +1,6 @@
+package de.maxhenkel.tools;
+
+@FunctionalInterface
+public interface FloatSupplier {
+    float getAsFloat();
+}